Update DSC.cpp
This is a major change, mainly to DSC.cpp.  Instead of recording an entire word as a binary string (which could use up to 80 or 100 bytes for just the initial word, then triple that for the old and new word strings, and then double all of that for the panel and keypad), it now stores the word as a byte array, mathematically summing the bits as they are received.  This saves an incredible amount of memory, now using less than 100 bytes for all of the data which would previously be stored using strings and up to 600 bytes.  Now an arduino uno will run this will also using the Ethernet library.  This combination used to cause a memory crash after a few seconds.